+++ .changeset/khaki-apricots-flow.md
@@ -1,0 +1,5 @@
+---
+"@thirdweb-dev/sdk": patch
+---
+
+Reduce circular dependencies
